+++ backend/src/main/java/za/co/rhome/config/ApplicationConfig.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@Configuration
 @ConfigurationProperties(prefix = "app")
 public class ApplicationConfig {
